New way of creating aliases in trial.php
Aliases are created for every column in the order file.
Headers are formatted to variables names using camelCase.

This means that "MinTime" no longer works, and "Min Time" is the proper header to create the $minTime variable.

For post trials, the aliases are made from the columns starting with "PostX", where X is the number of the post-trial that we are currently on.
After these aliases are created, any columns for the original trial that don't have new values for this post-trial are re-used.  For example, if there is no "Post1 Min Time", but there is a "Min Time", then the post trial will use "Min Time" as its $minTime.