Added ability to flag conditions, changed condition selection logic
closes #65

Flagging conditions is done so by putting a # at the start of the condition description

flagged conditions will not be selected when the selection is set to "auto"

If condition options are shown, flagged conditions are marked by grey text. However, they can be hidden completely by using the Settings.php file with the setting $hideFlaggedConditions
If flagged conditions are not hidden, they can still be selected and run, for running demos or special conditions like that.

Also, there is a logic change to how condition choices are submitted to login.php:
Instead of sending the column entry of "Condition Number", the index of the getFromFile() array is used, so that it doesn't matter if condition numbers are marked out sequentially. This sort of makes the error checking for numbers at the top of index.php unnecessary.