Update CustomFunctions.php
Adds PHPDoc style documentation to all the functions which is helpful when using an IDE (most will suggest the functions now and show the descriptions and arguments). 

Breaks up some of the larger functions to be under about 20 lines (excepting BlockShuffle, FileExists, and ComputeString). This allows for easier understanding of the code which will help with maintaining and updating it in the future.

Some superfluous code was simply removed (e.g. arguments that are unused throughout the app, conditionals that could be solved with existing PHP functions).

Updated the code so that it (all?) follows the Zend Coding conventions. This was just to make it a bit more readable and uniform, and only really applies to the formatting and spacing.