fixed possible bug in custom function fileExists()
before, it was possible that finding a directory match when $findDir was set to FALSE would result in the function returning FALSE, even if $findAltExtensions was set to TRUE and files with the same name, but also an extension, existed. For example, if the contents of the dir were test/ and test.file, and you used fileExists('test', true, false), it would return FALSE rather than continuing to search for test.file. This should be corrected.