Fixes the memory error with phpbrowscap
The Browscap project has three levels of ini files (full, normal, lite)
and phpbrowscap was using the largest of these three. Unfortunately it
is too large (and has a lot of unneccessary User Agents for our
purposes) so the Browscap code was updated to update using the Lite
version, and I primed the cache with the most current Lite version.

The code putting the UA info into the output files has been updated.
Instead of a long UA string in one column, there are three columns:
Browser, which will have the browser and version number; Device type,
which indicates whether it is desktop, tablet, or mobile; and OS, which
indicates the operating system.